wip(core): move search manager logic
