+++ src/appManager.ts
@@ -1,5 +1,5 @@
 import { getPublicTimeline, getHomeTimeline } from './api/timeline'
-import { type Search, type Status } from './types/shared'
+import { type Status } from './types/shared'
 import type { AppConfig } from './core/config'
 import { logOut, isLoaded as isUserLoaded } from './core/user'
 import { authorize } from './core/auth.ts'
@@ -7,11 +7,8 @@ import type { Router } from './router'
 import type { GlobalNavigation, ITimelineManager } from './types/shared'
 import { appConfig } from './core/config'
 import { lRouter } from './router'
-import { fetchJson } from './utils/fetch.ts'
 import { PageHistoryManager, usePageHistory } from './utils/pageHistory.ts'
-import { searchParams } from './utils/url.ts'
 import { last } from './utils/arrays.ts'
-import { createSignal } from './utils/signal.ts'
 
 export class TimelineManager implements ITimelineManager {
   private maxId: string
@@ -106,76 +103,16 @@ export function useGlobalNavigation(
   }
 }
 
-type SearchParams = {
-  q: string
-  limit?: string
-  type?: string
-}
-
-function createSearchManager() {
-  let res: Search,
-    offset = 0,
-    _noMoreData = false,
-    loading = createSignal(false)
-
-  let { server } = appConfig
-
-  function resetOffset() {
-    offset = 0
-    _noMoreData = false
-  }
-
-  async function search(opts: SearchParams) {
-    let q = searchParams({
-      ...opts,
-      offset: offset.toString(),
-    })
-
-    try {
-      loading(true)
-      res = await fetchJson(`${server()}/api/v2/search?${q}`, {
-        withCredentials: true,
-      })
-
-      let len = res.statuses.length
-      _noMoreData = len === 0
-      offset += len
-    } catch {
-      res = {
-        accounts: [],
-        statuses: [],
-        hashtags: [],
-      }
-    } finally {
-      loading(false)
-    }
-  }
-
-  return {
-    search,
-    get searchResult() {
-      return res
-    },
-    get noMoreData() {
-      return _noMoreData
-    },
-    resetOffset,
-    loading,
-  }
-}
-
 export class AppManager {
   private config: AppConfig
   public timelineManager: TimelineManager
   public globalMediator: GlobalNavigation
   public pageHistoryManager: PageHistoryManager
-  public searchManager: ReturnType<typeof createSearchManager>
 
   constructor() {
     this.config = appConfig
     this.timelineManager = new TimelineManager()
     this.pageHistoryManager = usePageHistory()
-    this.searchManager = createSearchManager()
 
     this.globalMediator = useGlobalNavigation(
       this.timelineManager,
